User models moved to users/model package . Now UserEntity is not an entity . it is a template for command requests .
