--- messenger-core/src/main/java/com/example/messenger/user/UserDeleteDto.java
@@ -1,21 +1,0 @@
-package com.example.messenger.user;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-import java.io.Serializable;
-
-/**
- * DTO for {@link UserEntity}
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Accessors(chain = true)
-public class UserDeleteDto implements Serializable {
-    @NotBlank(message = "Username cannot be null")
-    private String username;
-}
