--- messenger-core/src/main/java/com/example/messenger/user/UserDto.java
@@ -1,20 +1,0 @@
-package com.example.messenger.user;
-
-import lombok.Value;
-
-import java.util.UUID;
-
-/**
- * DTO for {@link UserEntity}
- */
-
-@Value
-public class UserDto {
-
-    UUID id;
-    String username;
-    String email;
-    String firstName;
-    String lastName;
-    String authorities;
-}
