+++ messenger-core/src/main/java/com/example/messenger/user/model/UserDeleteDto.java
@@ -1,0 +1,23 @@
+package com.example.messenger.user.model;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+
+/**
+ * DTO for {@link UserEntity}
+ */
+@Setter
+@Getter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@Accessors(chain = true)
+public class UserDeleteDto implements Serializable {
+    @NotNull(message = "Username cannot be null")
+    @NotEmpty(message = "Username cannot be empty")
+    private String username;
+}
